+++ CHANGELOG.md
@@ -1,5 +1,28 @@
 # Changelog
 
+## [0.4.1](https://github.com/codehausau/takcli/compare/takcli-v0.4.0...takcli-v0.4.1) (2026-03-19)
+
+
+### Bug Fixes
+
+* default deploy image tag to latest ([6f39e29](https://github.com/codehausau/takcli/commit/6f39e29b4e55ff078170d86bfc173402fec8ed5b))
+* validate deploy certificate passwords ([08af270](https://github.com/codehausau/takcli/commit/08af2705292a1e3147a717f222190012f30af6f2))
+* validate deploy passwords and default image tag to latest ([44269d0](https://github.com/codehausau/takcli/commit/44269d085dc4cf88d468c792785fd24dd3fcdc89))
+
+## [0.4.0](https://github.com/codehausau/takcli/compare/takcli-v0.3.0...takcli-v0.4.0) (2026-03-19)
+
+
+### Features
+
+* add compose deployment workflow ([748c3d3](https://github.com/codehausau/takcli/commit/748c3d3bc2975c54e176f551b4b84e1a8bf7dcf5))
+* deploy compose workflow ([6db1eab](https://github.com/codehausau/takcli/commit/6db1eabe0d173f37f88fa7101b96c112363e8dd0))
+* prompt for deploy secrets ([1e9223c](https://github.com/codehausau/takcli/commit/1e9223c7a58227b7bbbd94574bc5b7af1390b080))
+
+
+### Bug Fixes
+
+* retry transient cot query timeouts ([bbce2a0](https://github.com/codehausau/takcli/commit/bbce2a050afb610db85b05f07cf5ae9c5d3331c6))
+
 ## [0.3.0](https://github.com/codehausau/takcli/compare/takcli-v0.2.0...takcli-v0.3.0) (2026-03-17)